% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_list.R
\name{download_list}
\alias{download_list}
\title{Download List}
\usage{
download_list()
}
\arguments{
\item{none}{none}
}
\description{
Downloads the players list in XML format from Fide
}
\examples{
download_list()
}
\keyword{Fide}
\keyword{download}
\keyword{list}
\keyword{players}

