% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_xml.R
\name{process_xml}
\alias{process_xml}
\title{Process XML}
\usage{
process_xml(date)
}
\arguments{
\item{date}{download date in dd_mm_yyyy format}
}
\description{
Processes the players list in XML format from Fide
}
\examples{
process_xml("02_04_2016")
}
\keyword{Fide}
\keyword{XML}
\keyword{list}
\keyword{players}
\keyword{process}

